requirements.txt
fastapi[standard]
Jinja2
asyncpg
python-dotenv
SQLAlchemy
psycopg2-binary

annotated-doc
annotated-types
anyio